requirements.txt
coverage==4.4.1
Django==1.11.1
django-filter==1.0.4
djangorestframework==3.6.3
Markdown==2.6.8
psycopg2==2.7.1
pytz==2017.2
